requirements.txt
xlsxwriter>=3.1.1
google-genai>=0.4.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
Flask>=3.0.0
Flask-Mail>=0.9.1
croniter>=2.0.5
